Ajout squelette projet

+++ td4/src/MySymbol.java
@@ -5,7 +5,7 @@ public class MySymbol extends Symbol {
     final public int value;
 
     public MySymbol(int value) {
-        super((short)0);
+        super(0);
         this.value = value;
     }
